No more InversionFlowlines
